Toggle between filters/search results on mobile
This uses a clever hack around how the CSS `hover` state is implemented on
mobile devices. Typically, the `hover` state is only shown when the user clicks
on the object, but after that click it _stays_ hovered until something else is
clicked.

We cram the filters into a small div with a title that, when hover-clicked,
expands to take up the entire section, but leaves a search-results header div
visible. When that's clicked, it de-hovers the filters div and returns the view
to the search results.